VADER & Textblob
Rule-based approaches rely on a predefined set of rules, patterns, and dictionaries to identify the sentiment of the text. These rules typically include the list of positive, neutral, and negative keywords, grammatical structures, and linguistic patterns.
VADER and Textblob was implemented. Preprocessing of text was excluded as it has resulted in poor performance. After the implementation of both the methods, it was found that the output was not 100% accurate. 
The limitations of VADER are It usually has a hard time understanding Sarcasm and Irony. Also, negations are not easily identified.